+++ packages/eslint-config/index.js
@@ -7,6 +7,8 @@ module.exports = {
     'eslint-config-prettier/react',
   ].map(require.resolve),
 
+  plugins: ['react-hooks'],
+
   parserOptions: {
     ecmaVersion: 7,
     sourceType: 'module',
@@ -86,11 +88,20 @@ module.exports = {
     ],
     'jsx-a11y/label-has-associated-control': 0,
     'jsx-a11y/label-has-for': 0,
+    'react-hooks/rules-of-hooks': 'error',
+    'react-hooks/exhaustive-deps': 'warn',
   },
   overrides: [
     {
-      files: ['*.json'],
+      files: ['**/*.json'],
       parser: 'json',
+    },
+    {
+      parser: require.resolve('@typescript-eslint/parser'),
+      files: ['**/*.ts', '**/*.tsx'],
+      rules: {
+        'no-unused-vars': 0,
+      },
     },
   ],
 };