+++ packages/gamut/Icon/iconMap.tsx
@@ -72,6 +72,9 @@ import UpArrowIcon from './icons/UpArrowIcon';
 import VideoIcon from './icons/VideoIcon';
 import jQueryIcon from './icons/jQueryIcon';
 
+/**
+ * @deprecated Directly import icons from @codecademy/gamut instead.
+ */
 const iconMap = {
   accessibility: AccessibilityIcon,
   add: AddIcon,
@@ -149,4 +152,7 @@ const iconMap = {
   video: VideoIcon,
 };
 
+/**
+ * @deprecated Directly import icons from @codecademy/gamut instead.
+ */
 export default iconMap;