+++ packages/gamut/Icon/icons/AccessibilityIcon.tsx
@@ -1,6 +1,8 @@
-import React, { SVGProps } from 'react';
+import React from 'react';
 
-export default function AccessibilityIcon(props: SVGProps<SVGSVGElement>) {
+import { wrapIcon } from '../wrapIcon';
+
+export default wrapIcon(function AccessibilityIcon(props) {
   return (
     <svg viewBox="0 0 24 24" {...props}>
       <title>Accessibility Icon</title>
@@ -8,4 +10,4 @@ export default function AccessibilityIcon(props: SVGProps<SVGSVGElement>) {
       <path d="M11.966 5.81c-.841 0-1.527.68-1.527 1.515 0 .834.686 1.515 1.527 1.515.842 0 1.527-.68 1.527-1.515 0-.835-.685-1.515-1.527-1.515zM16.19 9.042c.266 0 .509.238.554.561.049.347-.146.663-.436.703l-3.398.473v2.113l1.833 4.36a.643.643 0 0 1-.594.88.649.649 0 0 1-.6-.394l-1.684-4.006-1.685 4.006a.649.649 0 0 1-.843.344.637.637 0 0 1-.35-.83l1.833-4.36V10.78l-3.399-.47c-.289-.04-.486-.357-.435-.704.045-.323.29-.561.554-.561.02 0 .04.002.06.004l3.282.456h1.965l3.282-.456a.282.282 0 0 1 .06-.006z" />
     </svg>
   );
-}
+});