+++ packages/gamut/Icon/icons/CheckFilledIcon.tsx
@@ -1,6 +1,8 @@
-import React, { SVGProps } from 'react';
+import React from 'react';
 
-export default function CheckFilledIcon(props: SVGProps<SVGSVGElement>) {
+import { wrapIcon } from '../wrapIcon';
+
+export default wrapIcon(function CheckFilledIcon(props) {
   return (
     <svg viewBox="0 0 24 24" {...props}>
       <title>CheckFilled Icon</title>
@@ -12,4 +14,4 @@ export default function CheckFilledIcon(props: SVGProps<SVGSVGElement>) {
       </g>
     </svg>
   );
-}
+});