+++ packages/gamut/Icon/icons/MatplotLibIcon.tsx
@@ -1,6 +1,8 @@
-import React, { SVGProps } from 'react';
+import React from 'react';
 
-export default function MatplotLibIcon(props: SVGProps<SVGSVGElement>) {
+import { wrapIcon } from '../wrapIcon';
+
+export default wrapIcon(function MatplotLibIcon(props) {
   return (
     <svg viewBox="0 0 24 24" {...props}>
       <title>MatplotLib Icon</title>
@@ -18,4 +20,4 @@ export default function MatplotLibIcon(props: SVGProps<SVGSVGElement>) {
       </g>
     </svg>
   );
-}
+});